+++ sarif/cmdline/main.py
@@ -3,11 +3,10 @@ Program entry point for sarif-tools on the command line.
 """
 
 import argparse
-from importlib import metadata
 import os
 import sys
 
-from sarif import loader, sarif_file
+from sarif import loader, sarif_file, __version__ as SARIF_TOOLS_PACKAGE_VERSION
 from sarif.filter.general_filter import load_filter_file
 
 from sarif.operations import (
@@ -34,12 +33,17 @@ def main():
     args, unknown_args = ARG_PARSER.parse_known_args()
 
     if args.debug:
-        print(f"SARIF tools v{SARIF_TOOLS_PACKAGE_VERSION}")
+        _print_version()
         print(f"Running code from {__file__}")
         known_args_summary = ", ".join(
             f"{key}={getattr(args, key)}" for key in vars(args)
         )
         print(f"Known arguments: {known_args_summary}")
+        if args.version:
+            return 0
+    elif args.version:
+        _print_version()
+        return 0
 
     if unknown_args:
         if any(
@@ -65,14 +69,13 @@ def _create_arg_parser():
     for cmd, cmd_attributes in _COMMANDS.items():
         cmd_list += cmd.ljust(col_width) + cmd_attributes["desc"] + "\n"
     cmd_list += "Run `sarif <COMMAND> --help` for command-specific help."
-    package_version = _read_package_version()
     parser = argparse.ArgumentParser(
         prog="sarif",
         description="Process sets of SARIF files",
         epilog=cmd_list,
         formatter_class=argparse.RawDescriptionHelpFormatter,
     )
-    parser.set_defaults(func=_usage)
+    parser.set_defaults(func=_usage_command)
     subparsers = parser.add_subparsers(dest="command", help="command")
     subparser = {}
     for cmd, cmd_attributes in _COMMANDS.items():
@@ -80,12 +83,7 @@ def _create_arg_parser():
         subparser[cmd].set_defaults(func=cmd_attributes["fn"])
 
     # Common options
-    parser.add_argument(
-        "--version",
-        "-v",
-        action="version",
-        version=f"sarif-tools version {package_version}",
-    )
+    parser.add_argument("--version", "-v", action="store_true")
     parser.add_argument(
         "--debug", action="store_true", help="Print information useful for debugging"
     )
@@ -93,7 +91,7 @@ def _create_arg_parser():
         "--check",
         "-x",
         type=str,
-        choices=sarif_file.SARIF_SEVERITIES,
+        choices=sarif_file.SARIF_SEVERITIES_WITH_NONE,
         help="Exit with error code if there are any issues of the specified level "
         + "(or for diff, an increase in issues at that level).",
     )
@@ -182,7 +180,7 @@ def _create_arg_parser():
         )
     # Most commands take an arbitrary list of SARIF files or directories
     for cmd in _COMMANDS:
-        if cmd not in ["diff", "upgrade-filter", "usage"]:
+        if cmd not in ["diff", "upgrade-filter", "usage", "version"]:
             subparser[cmd].add_argument(
                 "files_or_dirs",
                 metavar="file_or_dir",
@@ -225,18 +223,11 @@ def _create_arg_parser():
     return parser
 
 
-def _read_package_version():
-    try:
-        return metadata.version("sarif-tools")
-    except metadata.PackageNotFoundError:
-        return "local"
-
-
 def _check(input_files: sarif_file.SarifFileSet, check_level):
     ret = 0
     if check_level:
         counts = input_files.get_result_count_by_severity()
-        for severity in sarif_file.SARIF_SEVERITIES:
+        for severity in sarif_file.SARIF_SEVERITIES_WITH_NONE:
             ret += counts.get(severity, 0)
             if severity == check_level:
                 break
@@ -312,7 +303,7 @@ def _prepare_output(
 ####################################### Command handlers #######################################
 
 
-def _blame(args):
+def _blame_command(args):
     input_files = loader.load_sarif_files(*args.files_or_dirs)
     (output, multiple_file_output) = _prepare_output(input_files, args.output, ".sarif")
     blame_op.enhance_with_blame(
@@ -321,7 +312,7 @@ def _blame(args):
     return _check(input_files, args.check)
 
 
-def _codeclimate(args):
+def _codeclimate_command(args):
     input_files = loader.load_sarif_files(*args.files_or_dirs)
     input_files.init_default_line_number_1()
     _init_path_prefix_stripping(input_files, args, strip_by_default=False)
@@ -331,7 +322,7 @@ def _codeclimate(args):
     return _check(input_files, args.check)
 
 
-def _copy(args):
+def _copy_command(args):
     input_files = loader.load_sarif_files(*args.files_or_dirs)
     _init_filtering(input_files, args)
     output = args.output or "out.sarif"
@@ -345,7 +336,7 @@ def _copy(args):
     return _check(output_sarif_file_set, args.check)
 
 
-def _csv(args):
+def _csv_command(args):
     input_files = loader.load_sarif_files(*args.files_or_dirs)
     input_files.init_default_line_number_1()
     _init_path_prefix_stripping(input_files, args, strip_by_default=False)
@@ -355,7 +346,7 @@ def _csv(args):
     return _check(input_files, args.check)
 
 
-def _diff(args):
+def _diff_command(args):
     original_sarif = loader.load_sarif_files(args.old_file_or_dir[0])
     new_sarif = loader.load_sarif_files(args.new_file_or_dir[0])
     _init_filtering(original_sarif, args)
@@ -363,7 +354,7 @@ def _diff(args):
     return diff_op.print_diff(original_sarif, new_sarif, args.output, args.check)
 
 
-def _html(args):
+def _html_command(args):
     input_files = loader.load_sarif_files(*args.files_or_dirs)
     input_files.init_default_line_number_1()
     _init_path_prefix_stripping(input_files, args, strip_by_default=True)
@@ -373,7 +364,7 @@ def _html(args):
     return _check(input_files, args.check)
 
 
-def _emacs(args):
+def _emacs_command(args):
     input_files = loader.load_sarif_files(*args.files_or_dirs)
     input_files.init_default_line_number_1()
     _init_path_prefix_stripping(input_files, args, strip_by_default=True)
@@ -383,7 +374,7 @@ def _emacs(args):
     return _check(input_files, args.check)
 
 
-def _info(args):
+def _info_command(args):
     input_files = loader.load_sarif_files(*args.files_or_dirs)
     info_op.generate_info(input_files, args.output)
     if args.check:
@@ -391,7 +382,7 @@ def _info(args):
     return 0
 
 
-def _ls(args):
+def _ls_command(args):
     ls_op.print_ls(args.files_or_dirs, args.output)
     if args.check:
         input_files = loader.load_sarif_files(*args.files_or_dirs)
@@ -399,7 +390,7 @@ def _ls(args):
     return 0
 
 
-def _summary(args):
+def _summary_command(args):
     input_files = loader.load_sarif_files(*args.files_or_dirs)
     _init_filtering(input_files, args)
     (output, multiple_file_output) = (None, False)
@@ -411,7 +402,7 @@ def _summary(args):
     return _check(input_files, args.check)
 
 
-def _trend(args):
+def _trend_command(args):
     input_files = loader.load_sarif_files(*args.files_or_dirs)
     input_files.init_default_line_number_1()
     _init_filtering(input_files, args)
@@ -424,7 +415,7 @@ def _trend(args):
     return _check(input_files, args.check)
 
 
-def _upgrade_filter(args):
+def _upgrade_filter_command(args):
     old_filter_files = args.files_or_dirs
     single_output_file = None
     output_dir = None
@@ -445,7 +436,7 @@ def _upgrade_filter(args):
     return 0
 
 
-def _usage(args):
+def _usage_command(args):
     if hasattr(args, "output") and args.output:
         with open(args.output, "w", encoding="utf-8") as file_out:
             ARG_PARSER.print_help(file_out)
@@ -458,7 +449,19 @@ def _usage(args):
     return 0
 
 
-def _word(args):
+def _version_command(args):
+    _print_version(not args.version)
+
+
+def _print_version(bare=False):
+    print(
+        SARIF_TOOLS_PACKAGE_VERSION
+        if bare
+        else f"SARIF tools v{SARIF_TOOLS_PACKAGE_VERSION}"
+    )
+
+
+def _word_command(args):
     input_files = loader.load_sarif_files(*args.files_or_dirs)
     input_files.init_default_line_number_1()
     _init_path_prefix_stripping(input_files, args, strip_by_default=True)
@@ -472,55 +475,64 @@ def _word(args):
 
 _COMMANDS = {
     "blame": {
-        "fn": _blame,
+        "fn": _blame_command,
         "desc": "Enhance SARIF file with information from `git blame`",
     },
     "codeclimate": {
-        "fn": _codeclimate,
+        "fn": _codeclimate_command,
         "desc": "Write a JSON representation in Code Climate format of SARIF file(s) "
         "for viewing as a Code Quality report in GitLab UI",
     },
     "copy": {
-        "fn": _copy,
+        "fn": _copy_command,
         "desc": "Write a new SARIF file containing optionally-filtered data from other SARIF file(s)",
     },
     "csv": {
-        "fn": _csv,
+        "fn": _csv_command,
         "desc": "Write a CSV file listing the issues from the SARIF files(s) specified",
     },
     "diff": {
-        "fn": _diff,
+        "fn": _diff_command,
         "desc": "Find the difference between two [sets of] SARIF files",
     },
     "emacs": {
-        "fn": _emacs,
+        "fn": _emacs_command,
         "desc": "Write a representation of SARIF file(s) for viewing in emacs",
     },
     "html": {
-        "fn": _html,
+        "fn": _html_command,
         "desc": "Write an HTML representation of SARIF file(s) for viewing in a web browser",
     },
-    "info": {"fn": _info, "desc": "Print information about SARIF file(s) structure"},
-    "ls": {"fn": _ls, "desc": "List all SARIF files in the directories specified"},
+    "info": {
+        "fn": _info_command,
+        "desc": "Print information about SARIF file(s) structure",
+    },
+    "ls": {
+        "fn": _ls_command,
+        "desc": "List all SARIF files in the directories specified",
+    },
     "summary": {
-        "fn": _summary,
+        "fn": _summary_command,
         "desc": "Write a text summary with the counts of issues from the SARIF files(s) specified",
     },
     "trend": {
-        "fn": _trend,
+        "fn": _trend_command,
         "desc": "Write a CSV file with time series data from SARIF files with "
         '"yyyymmddThhmmssZ" timestamps in their filenames',
     },
     "upgrade-filter": {
-        "fn": _upgrade_filter,
+        "fn": _upgrade_filter_command,
         "desc": "Upgrade a sarif-tools v1-style blame filter file to a v2-style filter YAML file",
     },
-    "usage": {"fn": _usage, "desc": "(Command optional) - print usage and exit"},
+    "usage": {
+        "fn": _usage_command,
+        "desc": "(Command optional) - print usage and exit",
+    },
+    "version": {"fn": _version_command, "desc": "Print version and exit"},
     "word": {
-        "fn": _word,
+        "fn": _word_command,
         "desc": "Produce MS Word .docx summaries of the SARIF files specified",
     },
 }
 
-SARIF_TOOLS_PACKAGE_VERSION = _read_package_version()
 ARG_PARSER = _create_arg_parser()