+++ sarif/operations/csv_op.py
@@ -7,6 +7,7 @@ import os
 
 from sarif import sarif_file
 from sarif.sarif_file import SarifFileSet
+from sarif.sarif_file_utils import combine_record_code_and_description
 
 
 def generate_csv(input_files: SarifFileSet, output: str, output_multiple_files: bool):
@@ -47,7 +48,7 @@ def _write_to_csv(file_or_files, output_file):
     Write out the errors to a CSV file so that a human can do further analysis.
     """
     list_of_errors = file_or_files.get_records()
-    severities = sarif_file.SARIF_SEVERITIES
+    severities = file_or_files.get_severities()
     # newline="" to avoid \r\r\n - see https://stackoverflow.com/a/3191811/316578
     with open(output_file, "w", encoding="utf-8", newline="") as file_out:
         writer = csv.DictWriter(
@@ -59,6 +60,6 @@ def _write_to_csv(file_or_files, output_file):
                 e for e in list_of_errors if e["Severity"] == severity
             ]
             sorted_errors_by_severity = sorted(
-                errors_of_severity, key=sarif_file.combine_code_and_description
+                errors_of_severity, key=combine_record_code_and_description
             )
             writer.writerows(error_dict for error_dict in sorted_errors_by_severity)