+++ sarif/operations/diff_op.py
@@ -7,6 +7,7 @@ import sys
 from typing import Dict
 
 from sarif import sarif_file
+from sarif.sarif_file_utils import combine_record_code_and_description
 
 
 def _occurrences(occurrence_count):
@@ -41,7 +42,7 @@ def print_diff(
         with open(output, "w", encoding="utf-8") as output_file:
             json.dump(diff, output_file, indent=4)
     else:
-        for severity in sarif_file.SARIF_SEVERITIES:
+        for severity in sarif_file.SARIF_SEVERITIES_WITH_NONE:
             if diff[severity]["codes"]:
                 print(
                     severity,
@@ -86,7 +87,7 @@ def print_diff(
         print(f"  'After' results were filtered by {filter_stats}")
     ret = 0
     if check_level:
-        for severity in sarif_file.SARIF_SEVERITIES:
+        for severity in sarif_file.SARIF_SEVERITIES_WITH_NONE:
             ret += diff.get(severity, {}).get("+", 0)
             if severity == check_level:
                 break
@@ -101,12 +102,12 @@ def _find_new_occurrences(new_records, old_records, issue_code_and_desc):
     old_occurrences = [
         r
         for r in old_records
-        if sarif_file.combine_code_and_description(r) == issue_code_and_desc
+        if combine_record_code_and_description(r) == issue_code_and_desc
     ]
     new_occurrences_new_locations = []
     new_occurrences_new_lines = []
     for new_record in new_records:
-        if sarif_file.combine_code_and_description(new_record) == issue_code_and_desc:
+        if combine_record_code_and_description(new_record) == issue_code_and_desc:
             (new_location, new_line) = (True, True)
             for old_record in old_occurrences:
                 if old_record["Location"] == new_record["Location"]:
@@ -136,7 +137,7 @@ def calc_diff(
     Return dict has keys "error", "warning", "note" and "all".
     """
     ret = {"all": {"+": 0, "-": 0}}
-    for severity in sarif_file.SARIF_SEVERITIES:
+    for severity in sarif_file.SARIF_SEVERITIES_WITH_NONE:
         original_histogram = dict(original_sarif.get_issue_code_histogram(severity))
         new_histogram = new_sarif.get_issue_code_histogram(severity)
         new_histogram_dict = dict(new_histogram)